safari-600.7-branch/LayoutTests/http/tests/xmlhttprequest/upload-request-error-event-order-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/cross-site-progress-events.cgi. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
Test to validate the order in which the events are fired in case of a request error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS actualOrder is "rsdone,upload.onerror,upload.onloadend,onerror,onloadend,"
PASS successfullyParsed is true

TEST COMPLETE
